move renderDials into GridList component
